fixed entity autoincrement; debugging adding component to entity

+++ component.go
@@ -7,18 +7,18 @@ import (
 )
 
 type Component struct {
-	ID         int      `storm:"id,increment" json:"id"`
-	Name       string   `json:"name"`
-	Shape      string   `json:"shape"`
-	Material   string   `json:"material"`
-	Color      int       `json:"color"`
-	Size			 []float64 `json:"size"`
-	Position   []int    `json:"position"`
-	Quaternion []int    `json:"quaternion"`
-	ComponentType    string `json:"type"`
+	ID            int      `storm:"id,increment" json:"id"`
+	Name          string   `json:"name"`
+	Shape         string   `json:"shape"`
+	Material      string   `json:"material"`
+	Color         int       `json:"color"`
+	Size			    []float64 `json:"size"`
+	Position      []float64    `json:"position"`
+	Quaternion    []float64    `json:"quaternion"`
+	ComponentType string `json:"type"`
 }
 
-func NewComponent(name string, shape string, mat string, color int, size []float64, pos []int, quat []int, componentType string) *Component {
+func NewComponent(name string, shape string, mat string, color int, size []float64, pos []float64, quat []float64, componentType string) *Component {
 	return &Component{Name: name, Shape: shape, Material: mat, Color: color, Size: size, Position: pos, Quaternion: quat, ComponentType: componentType}
 }
